\name{Irwls-package}
\alias{Irwls-package}
\alias{Irwls}
\docType{package}
\title{
Iteratively re-weighted least squares using C++
}
\description{
As a demonstration of C++ classes exposed through the Rcpp package,
this package creates a reference class to generate repeated
least-squares fits.  The class is created with a response variable and
a fully-specified model matrix; fits are then generated with varying
weight vectors.  The design is that this package can be used by any
form of model-fitting that eventually uses re-weighted least squares.
}
\author{
John Chambers <jmc at r-project.org>
}
\keyword{ package }

